backend/requirements.txt
Flask==2.3.2
Flask-CORS==4.0.0
python-dotenv==1.0.0
python-telegram-bot==20.3
gunicorn==21.2.0
werkzeug==2.3.6